requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
h11==0.16.0
idna==3.10
jh2==5.0.9
jinja2==3.1.4
joblib==1.5.1
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.3
niquests==3.14.1
numpy==2.2.6
openmeteo-requests==1.5.0
openmeteo-sdk==1.20.0
orjson==3.10.18
pandas==2.2.3
pillow==11.0.0
pip==24.0
platformdirs==4.3.8
pydantic==2.11.5
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qh3==1.5.1
requests==2.32.3
requests-cache==1.2.1
retry-requests==2.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
threadpoolctl==3.6.0
torch==2.7.0
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
urllib3-future==2.12.922
uvicorn==0.34.3
wassima==1.2.2
